+++ src/main/java/com/besson/endfield/block/ModBlocks.java
@@ -35,6 +35,11 @@ public class ModBlocks {
 //    public static final Block BIG_STORAGE = registerBlocks("big_storage",
 //            new BigStorageBlock(AbstractBlock.Settings.create().strength(3f, 5f).nonOpaque()));
 
+    public static final Block SUPPLY_TERMINAL = registerBlocks("supply_terminal",
+            new SupplyTerminalBlock(AbstractBlock.Settings.create().strength(1f, 5f).nonOpaque()));
+    public static final Block GUN_TOWER = registerBlocksWithoutItem("gun_tower",
+            new GunTowerBlock(AbstractBlock.Settings.create().strength(3f, 5f).nonOpaque()));
+
     public static final Block REFINING_UNIT = registerBlocksWithoutItem("refining_unit",
             new RefiningUnitBlock(AbstractBlock.Settings.create().strength(3f, 5f).nonOpaque()));
     public static final Block REFINING_UNIT_SIDE = registerBlocks("refining_unit_side",