+++ src/main/java/com/besson/endfield/block/custom/GunTowerBlock.java
@@ -1,0 +1,29 @@
+package com.besson.endfield.block.custom;
+
+import com.besson.endfield.block.ModBlockEntityWithFacing;
+import com.besson.endfield.blockentity.ModBlockEntities;
+import com.besson.endfield.blockentity.custom.GunTowerBlockEntity;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.block.entity.BlockEntityTicker;
+import net.minecraft.block.entity.BlockEntityType;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+import org.jetbrains.annotations.Nullable;
+
+public class GunTowerBlock extends ModBlockEntityWithFacing {
+    public GunTowerBlock(Settings settings) {
+        super(settings);
+    }
+
+    @Override
+    public @Nullable BlockEntity createBlockEntity(BlockPos pos, BlockState state) {
+        return new GunTowerBlockEntity(pos, state);
+    }
+
+    @Override
+    public @Nullable <T extends BlockEntity> BlockEntityTicker<T> getTicker(World world, BlockState state, BlockEntityType<T> type) {
+        return checkType(type, ModBlockEntities.GUN_TOWER, ((world1, pos, state1, blockEntity) ->
+                GunTowerBlockEntity.tick(world1, pos, state1, (GunTowerBlockEntity) blockEntity)));
+    }
+}