+++ src/main/java/com/besson/endfield/datagen/ModLootTablesProvider.java
@@ -31,6 +31,8 @@ public class ModLootTablesProvider extends FabricBlockLootTableProvider {
         addDrop(ModBlocks.THERMAL_BANK);
         addDrop(ModBlocks.FLUID_PUMP);
         addDrop(ModBlocks.PROTOCOL_ANCHOR_CORE_PORT);
+        addDrop(ModBlocks.SUPPLY_TERMINAL);
+        addDrop(ModBlocks.GUN_TOWER);
 
         addDrop(ModBlocks.AMETHYST_MINERAL_VEIN_BLOCK);
         addDrop(ModBlocks.COAL_MINERAL_VEIN_BLOCK);