+++ src/main/java/com/besson/endfield/item/ModItems.java
@@ -49,6 +49,8 @@ public class ModItems {
             new ThermalBankItem(ModBlocks.THERMAL_BANK, new Item.Settings().rarity(Rarity.UNCOMMON)), ModBlocks.THERMAL_BANK_SIDE);
     public static final Item FLUID_PUMP_ITEM = registerItems("fluid_pump",
             new FluidPumpItem(ModBlocks.FLUID_PUMP, new Item.Settings().rarity(Rarity.RARE)));
+    public static final Item GUN_TOWER_ITEM = registerItems("gun_tower",
+            new GunTowerItem(ModBlocks.GUN_TOWER, new Item.Settings()));
 
     public static final Item AGGAGRIT = registerItems("aggagrit",new Item(new Item.Settings()));
     public static final Item AGGAGRIT_BLOCK = registerItems("aggagrit_block",new Item(new Item.Settings().rarity(Rarity.UNCOMMON)));