+++ src/main/java/com/besson/endfield/screen/ModScreens.java
@@ -80,6 +80,11 @@ public class ModScreens {
     public static final ScreenHandlerType<BigStorageScreenHandler> BIG_STORAGE_SCREEN =
             Registry.register(Registries.SCREEN_HANDLER, new Identifier(ArknightsEndfield.MOD_ID, "big_storage_screen"),
                     new ExtendedScreenHandlerType<>(BigStorageScreenHandler::new));
+
+    public static final ScreenHandlerType<SupplyTerminalScreenHandler> SUPPLY_TERMINAL_SCREEN =
+            Registry.register(Registries.SCREEN_HANDLER, new Identifier(ArknightsEndfield.MOD_ID, "supply_terminal_screen"),
+                    new ExtendedScreenHandlerType<>(SupplyTerminalScreenHandler::new));
+
     public static void register() {
 
     }