+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # 更新日志 / Changelog
+## 2025-11-09
+- 修正配方问题
+- Fixed recipe issues
+
+## 2025-11-03 -> 11-06
+- 测试验证`供货终端`
+- Tested and verified `Supply Terminal`
+- 测试验证`炮塔`
+- Tested and verified `Gun Tower`
+
 ## 2025-11-02
 - 调整了矿物生成的概率
 - Adjusted the probability of mineral generation